tests/submission/W3C/domhtml/HTMLDocument05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#dom-document-body" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<h1>This is a Document</h1>
<p>This is a paragraph.</p>
<img alt="Yellow image" height="50" src="/support/images/yellow-rectangle.png" width="100">
<p class="copyright"><a href="http://www.csail.mit.edu/"><acronym title="Massachusetts Institute of Technology">MIT</acronym></a>, <a href="http://www.ercim.eu/"><acronym title="European Research Consortium for Informatics and Mathematics">ERCIM</acronym></a>,
<a href="http://www.keio.ac.jp/">Keio</a></p>


<script>
var domtest = async_test("The body attribute is the element that contains the content for the     document.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vbody;
      var vid;
      var doc;
      doc = window.document;
      vbody = doc.body;

      vid = vbody.id;

      assert_equals(vid,"TEST-BODY");
       
    domtest.done();
}));

</script>
<p>The body attribute is the element that contains the content for the
    document.

    Retrieve the body attribute and examine its value for the id attribute.</p>
<div id="log"></div>
</body></html>
